feat: add context and context provider
